requirements.txt
pypdf 
gradio 
python-dotenv
crewai